[Refactor] Drop _chunkwise suffix from deltanet, gated_deltanet, gla
Rename kernels/, ops/, tests/, benchmarks/, and workloads/ files:
- deltanet_chunkwise → deltanet
- gated_deltanet_chunkwise → gated_deltanet
- gla_chunkwise → gla

All internal imports updated. 2361 tests pass, 0 failures.
