+++ tileops/ops/gla.py
@@ -2,7 +2,7 @@ from typing import Dict, Optional, Tuple
 
 import torch
 
-from tileops.kernels.gla_chunkwise import GLABwdKernel, GLAFwdKernel
+from tileops.kernels.gla import GLABwdKernel, GLAFwdKernel
 from tileops.kernels.kernel import Kernel
 
 from .op import Op
