+++ workloads/ops/deltanet_fwd.py
